+++ socialhome/streams/tests/test_tasks.py
@@ -1,27 +1,23 @@
-from datetime import datetime, timedelta
+from datetime import timedelta
 from unittest.mock import Mock
 
 from django.test.utils import override_settings
 from django.utils.timezone import now
-from freezegun import freeze_time
 
-from socialhome.streams.tasks import streams_tasks, groom_redis_precaches
+from socialhome.streams.tasks import streams_tasks, groom_redis_precaches, delete_redis_keys
 from socialhome.tests.utils import SocialhomeTestCase
 from socialhome.users.models import User
 from socialhome.users.tests.factories import UserFactory
 from socialhome.utils import get_redis_connection
 
 
-@freeze_time("2016-04-02")
 def test_streams_tasks():
     mock_scheduler = Mock()
     streams_tasks(mock_scheduler)
-    mock_scheduler.schedule.assert_called_once_with(
-        scheduled_time=datetime.utcnow(),
-        func=groom_redis_precaches,
-        interval=60*60*3,
-        timeout=60*60*2,
-    )
+    _args, kwargs = mock_scheduler.schedule.call_args_list[0]
+    assert kwargs["func"] == delete_redis_keys
+    _args, kwargs = mock_scheduler.schedule.call_args_list[1]
+    assert kwargs["func"] == groom_redis_precaches
 
 
 @override_settings(